// apps/frontend-vue2/src/modules/segment/dtos/segment.dto.ts
export class SegmentDto {
  id?: number;
  accountId?: number;
  name: string;
  description: string;
  query: string;
  recurrence: number;
  scheduleCloudTaskId: string;
  type: string;
  steps?: any;
  contactsLimit?: number | null;
  lastCount: number;
  lastCountEmail: number;
  lastCountEmailPercentage: number;
  lastCountWebPush: number;
  lastCountWebPushPercentage: number;
  lastCountMobilePush: number;
  lastCountMobilePushPercentage: number;
  lastCountPhone: number;
  lastCountPhonePercentage: number;
  lastCountWhatsapp: number;
  lastCountWhatsappPercentage: number;
  segmentInfo?: any;
  addUnsubscribed: boolean;
  addInvalid: boolean;
  addBounced: boolean;
  isRealTimeSegment: boolean;
  createdAt: Date;
  updatedAt: Date;
  countContacts?: number | string;
  status?: string;
  isProcessing?: boolean;
  hasFinishedProcessing?: boolean;
  isClickhouseSegment?: boolean;
  lastRunDate?: Date | undefined;

  constructor(segmentDto: SegmentDto = {} as SegmentDto) {
    this.id = segmentDto.id;
    this.accountId = segmentDto.accountId;
    this.name = segmentDto.name;
    this.description = segmentDto.description;
    this.query = segmentDto.query;
    this.recurrence = segmentDto.recurrence;
    this.scheduleCloudTaskId = segmentDto.scheduleCloudTaskId;
    this.type = segmentDto.type;
    this.steps = segmentDto.steps;
    this.contactsLimit = segmentDto.contactsLimit || null;
    this.lastCount = segmentDto.lastCount || 0;
    this.lastCountEmail = segmentDto.lastCountEmail || 0;
    this.lastCountEmailPercentage = segmentDto.lastCountEmailPercentage || 0;
    this.lastCountWebPush = segmentDto.lastCountWebPush || 0;
    this.lastCountWebPushPercentage = segmentDto.lastCountWebPushPercentage || 0;
    this.lastCountMobilePush = segmentDto.lastCountMobilePush || 0;
    this.lastCountMobilePushPercentage = segmentDto.lastCountMobilePushPercentage || 0;
    this.lastCountPhone = segmentDto.lastCountPhone || 0;
    this.lastCountPhonePercentage = segmentDto.lastCountPhonePercentage || 0;
    this.lastCountWhatsapp = segmentDto.lastCountWhatsapp || 0;
    this.lastCountWhatsappPercentage = segmentDto.lastCountWhatsappPercentage || 0;
    this.addUnsubscribed = segmentDto.addUnsubscribed;
    this.addInvalid = segmentDto.addInvalid;
    this.addBounced = segmentDto.addBounced;
    this.isRealTimeSegment = segmentDto.isRealTimeSegment;
    this.createdAt = segmentDto.createdAt;
    this.updatedAt = segmentDto.updatedAt;
    this.countContacts = segmentDto.countContacts || 0;
    this.segmentInfo = segmentDto.segmentInfo || [];
    this.status = segmentDto.status;
    this.isProcessing = segmentDto.isProcessing || false;
    this.hasFinishedProcessing = segmentDto.hasFinishedProcessing || false;
    this.isClickhouseSegment = segmentDto.isClickhouseSegment || false;
    this.lastRunDate = segmentDto.lastRunDate;
  }
}